Rename log_tools to just log
